+++ src/robot_control/src/resources.cpp
@@ -1,10 +1,10 @@
 #include "../include/robot_control/resources.h"
 
-float Resources::angleSum(float angle_a, float angle_b){
+double Resources::angleSum(double angle_a, double angle_b){
 	angle_a = normalizeAngle(angle_a);
 	angle_b = normalizeAngle(angle_b);
 	
-	float sum = angle_a + angle_b;
+	double sum = angle_a + angle_b;
     
     sum = fmod(sum,2*M_PI);
 
@@ -15,9 +15,9 @@ float Resources::angleSum(float angle_a, float angle_b){
 }
 
 
-float Resources::angleDiff(float angle_a, float angle_b){
-    float diff1 = angle_a - angle_b;
-    float diff2 = diff1 - 2*M_PI;
+double Resources::angleDiff(double angle_a, double angle_b){
+    double diff1 = angle_a - angle_b;
+    double diff2 = diff1 - 2*M_PI;
 
     if(fabs(diff1) <= fabs(diff2))
         return diff1;
@@ -25,7 +25,7 @@ float Resources::angleDiff(float angle_a, float angle_b){
     return diff2;
 }
 
-float Resources::transformYaw(float yaw){
+double Resources::transformYaw(double yaw){
 	yaw = fmod(yaw, 2*M_PI);
 
 	if(yaw < 0)
@@ -34,7 +34,7 @@ float Resources::transformYaw(float yaw){
 	return yaw;
 }
 
-float Resources::normalizeAngle(float angle){
+double Resources::normalizeAngle(double angle){
 	if(angle < 0){
 		angle += (2*M_PI);
 	}
